% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_Rttest.R
\name{new_Rttest}
\alias{new_Rttest}
\title{new_Rttest}
\usage{
new_Rttest(v1, v2, alpha)
}
\arguments{
\item{v1}{First vector}

\item{v2}{Second vector}

\item{alpha}{Alpha level}
}
\value{
A data frame of the vectors, the alpha level, the confidence intervals, and the p-value of a t-test
}
\description{
new_Rttest
}
\examples{
\dontrun{new_Rttest(v1 = x, v2 = y, alpha = 0.05)}
}
