% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{plot}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{Object to plot}

\item{...}{Extra inputs}
}
\value{
the plot
}
\description{
plot
}
\examples{
\dontrun{plot(obj)}
}
