% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Rttest}
\alias{print.Rttest}
\title{print}
\usage{
\method{print}{Rttest}(x)
}
\arguments{
\item{x}{Object to print}
}
\value{
The data frame of the Rttest object
}
\description{
print
}
\examples{
\dontrun{print(obj)}
}
